rename classes and remove most example merchants
