--- Gemfile
@@ -1,8 +1,0 @@
-source 'https://rubygems.org'
-
-# Use Capistrano for deployment
-gem 'capistrano' , '~> 3.4.1'
-gem 'capistrano-passenger'
-gem 'capistrano-rvm'
-gem 'capistrano-bundler'
-gem 'capistrano-recipes'
