--- merchant_sdk/__init__.py
@@ -1,4 +1,0 @@
-from .MerchantBaseLogic import MerchantBaseLogic
-from .MerchantServer import MerchantServer
-
-__all__ = [MerchantBaseLogic, MerchantServer]
