+++ pricewars/__init__.py
@@ -1,0 +1,4 @@
+from .PricewarsMerchant import PricewarsMerchant
+from .MerchantServer import MerchantServer
+
+__all__ = [PricewarsMerchant, MerchantServer]
