+++ pricewars/models/Offer.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
-from merchant_sdk.models.PricewarsObject import PricewarsObject
-from merchant_sdk.models import Product
+from pricewars.models.PricewarsObject import PricewarsObject
+from pricewars.models import Product
 
 
 class Offer(PricewarsObject):
