+++ pricewars/models/PricewarsObject.py
