+++ pricewars/models/Product.py
